+++ core-bundle/src/Resources/contao/pages/PageError401.php
@@ -35,6 +35,13 @@ class PageError401 extends Frontend
 		$obj401 = $this->prepare($objRootPage);
 		$objPage = $obj401->loadDetails();
 
+		// Reset inherited cache timeouts (see #231)
+		if (!$objPage->includeCache)
+		{
+			$objPage->cache = 0;
+			$objPage->clientCache = 0;
+		}
+
 		/** @var PageRegular $objHandler */
 		$objHandler = new $GLOBALS['TL_PTY']['regular']();
 
@@ -56,6 +63,13 @@ class PageError401 extends Frontend
 
 		$obj401 = $this->prepare($objRootPage);
 		$objPage = $obj401->loadDetails();
+
+		// Reset inherited cache timeouts (see #231)
+		if (!$objPage->includeCache)
+		{
+			$objPage->cache = 0;
+			$objPage->clientCache = 0;
+		}
 
 		/** @var PageRegular $objHandler */
 		$objHandler = new $GLOBALS['TL_PTY']['regular']();